added ableton api exposure to groovy pages plus example groovy pages

+++ trunk/pages-0.2/src/org/monome/pages/ableton/AbletonLooper.java
@@ -1,16 +1,16 @@
 package org.monome.pages.ableton;
 
 public class AbletonLooper {
-	public final static float STATE_STOPPED = 0.0f;
-	public final static float STATE_RECORDING = 1.0f;
-	public final static float STATE_PLAYING = 2.0f;
-	public final static float STATE_OVERDUB = 3.0f;
-	private float state = 0.0f;
+	public final static int STATE_STOPPED = 0;
+	public final static int STATE_RECORDING = 1;
+	public final static int STATE_PLAYING = 2;
+	public final static int STATE_OVERDUB = 3;
+	private int state = 0;
 	
-	public void setState(float state) {
+	public void setState(int state) {
 		this.state = state;
 	}
-	public float getState() {
+	public int getState() {
 		return state;
 	}
 }
